% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipo_cambio.R
\name{get_divisas}
\alias{get_divisas}
\title{Listar divisas disponibles en la API del BCRA}
\usage{
get_divisas()
}
\value{
Un tibble con columnas \code{codigo} y \code{descripcion}.
}
\description{
Listar divisas disponibles en la API del BCRA
}
\examples{
\dontrun{
get_divisas()
}
}
